uppdate 1.1.1

+++ crates/sui-data-ingestion-core/src/reader.rs
@@ -75,7 +75,7 @@ impl CheckpointReader {
     /// Reads files in a local directory, validates them, and forwards `CheckpointData` to the executor.
     async fn read_local_files(&self) -> Result<Vec<Arc<CheckpointData>>> {
         let mut checkpoints = vec![];
-        for offset in 0 .. MAX_CHECKPOINTS_IN_PROGRESS {
+        for offset in 0 .. *MAX_CHECKPOINTS_IN_PROGRESS {
             let sequence_number = self.current_checkpoint_number + offset as u64;
             if self.exceeds_capacity(sequence_number) {
                 break;
@@ -92,11 +92,11 @@ impl CheckpointReader {
     }
 
     fn exceeds_capacity(&self, checkpoint_number: CheckpointSequenceNumber) -> bool {
-        ((MAX_CHECKPOINTS_IN_PROGRESS as u64 + self.last_pruned_watermark) <= checkpoint_number)
+        ((*MAX_CHECKPOINTS_IN_PROGRESS as u64 + self.last_pruned_watermark) <= checkpoint_number)
             || self.data_limiter.exceeds()
     }
 
-    async fn fetch_from_object_store(
+    pub async fn fetch_from_object_store(
         store: &dyn ObjectStore,
         checkpoint_number: CheckpointSequenceNumber,
     ) -> Result<(Arc<CheckpointData>, usize)> {
@@ -110,7 +110,7 @@ impl CheckpointReader {
         client: &Client,
         checkpoint_number: CheckpointSequenceNumber,
     ) -> Result<(Arc<CheckpointData>, usize)> {
-        let checkpoint = client.get_full_checkpoint(checkpoint_number).await?;
+        let checkpoint = client.clone().get_full_checkpoint(checkpoint_number).await?;
         let size = bcs::serialized_size(&checkpoint)?;
         Ok((Arc::new(checkpoint), size))
     }
@@ -134,14 +134,20 @@ impl CheckpointReader {
         checkpoint_number: CheckpointSequenceNumber,
     ) -> Result<(Arc<CheckpointData>, usize)> {
         let mut backoff = backoff::ExponentialBackoff::default();
-        backoff.max_elapsed_time = Some(Duration::from_secs(60));
+        let max_elapsed_time = Duration::from_secs(60);
+        backoff.max_elapsed_time = Some(max_elapsed_time);
         backoff.initial_interval = Duration::from_millis(100);
         backoff.current_interval = backoff.initial_interval;
         backoff.multiplier = 1.0;
         loop {
-            match Self::remote_fetch_checkpoint_internal(store, checkpoint_number).await {
-                Ok(data) => return Ok(data),
-                Err(err) => match backoff.next_backoff() {
+            match tokio::time::timeout(
+                max_elapsed_time,
+                Self::remote_fetch_checkpoint_internal(store, checkpoint_number),
+            )
+            .await
+            {
+                Ok(Ok(data)) => return Ok(data),
+                Ok(Err(err)) => match backoff.next_backoff() {
                     Some(duration) => {
                         if !err.to_string().contains("404") {
                             debug!("remote reader retry in {} ms. Error is {:?}", duration.as_millis(), err);
@@ -149,6 +155,10 @@ impl CheckpointReader {
                         tokio::time::sleep(duration).await
                     }
                     None => return Err(err),
+                },
+                Err(err) => match backoff.next_backoff() {
+                    Some(duration) => tokio::time::sleep(duration).await,
+                    None => return Err(err.into()),
                 },
             }
         }
@@ -270,10 +280,10 @@ impl CheckpointReader {
         for entry in fs::read_dir(self.path.clone())? {
             let entry = entry?;
             let filename = entry.file_name();
-            if let Some(sequence_number) = Self::checkpoint_number_from_file_path(&filename) {
-                if sequence_number < watermark {
-                    fs::remove_file(entry.path())?;
-                }
+            if let Some(sequence_number) = Self::checkpoint_number_from_file_path(&filename)
+                && sequence_number < watermark
+            {
+                fs::remove_file(entry.path())?;
             }
         }
         Ok(())
@@ -290,8 +300,8 @@ impl CheckpointReader {
         remote_store_options: Vec<(String, String)>,
         options: ReaderOptions,
     ) -> (Self, mpsc::Receiver<Arc<CheckpointData>>, mpsc::Sender<CheckpointSequenceNumber>, oneshot::Sender<()>) {
-        let (checkpoint_sender, checkpoint_recv) = mpsc::channel(MAX_CHECKPOINTS_IN_PROGRESS);
-        let (processed_sender, processed_receiver) = mpsc::channel(MAX_CHECKPOINTS_IN_PROGRESS);
+        let (checkpoint_sender, checkpoint_recv) = mpsc::channel(*MAX_CHECKPOINTS_IN_PROGRESS);
+        let (processed_sender, processed_receiver) = mpsc::channel(*MAX_CHECKPOINTS_IN_PROGRESS);
         let (exit_sender, exit_receiver) = oneshot::channel();
         let reader = Self {
             path,
